Fixed the tileset sizing issue
The Sunnyland art assets are now properly 16x16 pixels instead of 100 - this means basically everything in the environment broke though. I've worked on fixing pages 1 and 2 but the rest are still in disarray. Lots of little problems still - the player physics are wonky and the scroll speed in the scrolling-around view is super slow now. Spike walls are in random locations, the player spawn location is off - lots to do still