Added a lever that enables/disables SwitchBlocks
I added a lever that makes a sound effect and enables or disables the SwitchBlock objects in the scene when the player touches it. I also reduced the volume of wallsliding because it was annoying, and I changed a single typo in the former 'CharacterConroller' folder and changed it to 'CharacterController' which is why all the Bookwyrm stuff says it was affected. I didn't touch it otherwise.